switching to gradle, some checkstyle fixes, refactoring to org.xbib.jdbc

+++ src/main/java/org/xbib/jdbc/csv/BetweenExpression.java
@@ -1,21 +1,3 @@
-/*
- *  CsvJdbc - a JDBC driver for CSV files
- *  Copyright (C) 2008  Mario Frasca
- *
- *  This library is free software; you can redistribute it and/or
- *  modify it under the terms of the GNU Lesser General Public
- *  License as published by the Free Software Foundation; either
- *  version 2.1 of the License, or (at your option) any later version.
- *
- *  This library is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- *  Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public
- *  License along with this library; if not, write to the Free Software
- *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- */
 package org.xbib.jdbc.csv;
 
 import java.sql.SQLException;
@@ -23,6 +5,9 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
+/**
+ *
+ */
 class BetweenExpression extends LogicalExpression {
     Expression obj, left, right;
 
@@ -38,9 +23,9 @@ class BetweenExpression extends LogicalExpression {
         Comparable objValue = (Comparable) obj.eval(env);
         Integer comparedLeft = RelopExpression.compare(leftValue, objValue, env);
         boolean result = false;
-        if (comparedLeft != null && comparedLeft.intValue() <= 0) {
+        if (comparedLeft != null && comparedLeft <= 0) {
             Integer comparedRight = RelopExpression.compare(rightValue, objValue, env);
-            if (comparedRight != null && comparedRight.intValue() >= 0) {
+            if (comparedRight != null && comparedRight >= 0) {
                 result = true;
             }
         }
